% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past1.R
\name{past1}
\alias{past1}
\title{merge multiple elements splitted by a specific symbol}
\usage{
past1(x, y, split)
}
\arguments{
\item{x}{a character vector with one element}

\item{y}{a character vector with one element}

\item{split}{what to split on}
}
\value{
a character vector
}
\description{
merge multiple elements splitted by a specific symbol
}
\examples{
x <- "asfe"
y <- "434"
past1(x,y,split=".")
}
